Added route for deleting vehicles. Fixed cart bugs. - Modified database fields to support more precise values; - Bumped Prisma to newer version.

+++ server/prisma/db.sql
@@ -104,7 +104,7 @@ CREATE TABLE Address (
 CREATE TABLE Warehouse (
     id              INT UNSIGNED AUTO_INCREMENT,
     address         INT UNSIGNED NOT NULL,
-    capacity        INT UNSIGNED NOT NULL,
+    capacity        NUMERIC(10, 2) UNSIGNED NOT NULL,
     resource_usage  NUMERIC(10, 2) UNSIGNED NOT NULL,
     supplier        INT UNSIGNED NOT NULL,
     renewable_resources INT UNSIGNED NOT NULL,
@@ -124,7 +124,7 @@ CREATE TABLE Warehouse (
 CREATE TABLE Distribution_Center (
     id          INT UNSIGNED AUTO_INCREMENT,
     address     INT UNSIGNED NOT NULL,
-    capacity    INT UNSIGNED NOT NULL,
+    capacity    NUMERIC(10, 2) UNSIGNED NOT NULL,
     transporter INT UNSIGNED NOT NULL,
 
     PRIMARY KEY (id, transporter), # Composite PK allows dist. center indexing per transporter.
